SMPDSv2 dataset

SPECIAL Modern Pollen Data for climate reconstructionS, version 2.

SMPDSv2

Format

A data frame (tibble object) with 23843 rows and 2812 variables:

...

int ...

References

Davis BA, Chevalier M, Sommer P, Carter VA, Finsinger W, Mauri A, Phelps LN, Zanon M, Abegglen R, CM, others (2020). “The Eurasian Modern Pollen Database (EMPD), version 2.” Earth system science data, 12(4), 2423--2445. doi: 10.5194/essd-12-2423-2020 .

Finsinger W, Heiri O, Valsecchi V, Tinner W, Lotter AF (2007). “Modern pollen assemblages as climate indicators in southern Europe.” Global Ecology and Biogeography, 16(5), 567--582. doi: 10.1111/j.1466-8238.2007.00313.x .

Harrison SP (2019). “Modern pollen data for climate reconstructions, version 1 (SMPDS).” doi: 10.17864/1947.194 , https://researchdata.reading.ac.uk/194/.

Harrison SP, Shen Y, Sweeney L (2021). “Pollen data and charcoal data of the Iberian Peninsula.” doi: 10.17864/1947.294 .

Herzschuh U, Cao X, Laepple T, Dallmeyer A, Telford RJ, Ni J, Chen F, Kong Z, Liu G, Liu K, others (2019). “Position and orientation of the westerly jet determined Holocene rainfall patterns in China.” Nature communications, 10(1), 1--8. doi: 10.1038/s41467-019-09866-8 .

Phelps LN, Chevalier M, Shanahan TM, Aleman JC, Courtney-Mustaphi C, Kiahtipes CA, Broennimann O, Marchant R, Shekeine J, Quick LJ, others (2020). “Asymmetric response of forest and grassy biomes to climate variability across the African Humid Period: influenced by anthropogenic disturbance?” Ecography, 43(8), 1118--1142. doi: 10.1111/ecog.04990 .

Vincens A, Lézine A, Buchet G, Lewden D, Le Thomas A (2007). “African pollen database inventory of tree and shrub pollen types.” Review of Palaeobotany and Palynology, 145(1-2), 135--141. doi: 10.1016/j.revpalbo.2006.09.004 .

Williams JW, Grimm EC, Blois JL, Charles DF, Davis EB, Goring SJ, Graham RW, Smith AJ, Anderson M, Arroyo-Cabrales J, others (2018). “The Neotoma Paleoecology Database, a multiproxy, international, community-curated data resource.” Quaternary Research, 89(1), 156--177. doi: 10.1017/qua.2017.105 .