Compare two datasets using latitude and longitude

Compare two datasets using latitude and longitude, the comparison is done using the function fx, this can be a custom function or an existing function like round, trunc, etc.

compare_latlon(x, y, join_method = dplyr::inner_join, fx = tolerance, ...)

Arguments

x

Reference dataset.

y

Target dataset, this will be compared with x.

join_method

Function to be used as the joining method between x and y. Default: dplyr::inner_join.

fx

Comparison function, it defaults to an internal function called tolerance that the drops decimals (using the parameter digits) without rounding the numbers. Suggested base functions are round, trunc, floor, ceiling and signif.

...

Optional parameters passed to fx.

Value

Table with records of y that exist in x.