Calculate MI

Calculate Moisture Index (MI) from daily values of precipitation, sunshine fraction and temperature.

mi(.data, ...)

# S3 method for tbl_df
mi(.data, cpus = 1, ...)

Arguments

.data

Data frame (tibble object) (see the details section).

...

Optional parameters, including cpus.

cpus

Numeric value with the number of CPUs to be used for the computation. Default: 1, serial computation.

Value

Data frame (tibble object) with the same input data and the addition of a column called mi, containing values for MI.

Details

The input (.data object) data, should contain the following columns:

latitude: numeric values for latitude (decimal degrees).

elevation: numeric values for elevation (m A.S.L.)

pre: list of numeric values for daily precipitation (mm / day).

sf: list of numeric values for daily sunshine fraction (%).

tmp: list of numeric values for daily temperature (degrees Celsius).

See also